+++ src/clients/api/queries/useGetVaults/useGetXvsVaultPoolBalances.ts
@@ -2,10 +2,10 @@ import { useQueries, UseQueryOptions, UseQueryResult } from 'react-query';
 
 import { useWeb3 } from 'clients/web3';
 import { getTokenContractByAddress } from 'clients/contracts';
-import { getBalanceOf, GetBalanceOfOutput } from 'clients/api';
 import FunctionKey from 'constants/functionKey';
 import { Bep20 } from 'types/contracts';
-import { XVS_VAULT_PROXY_CONTRACT_ADDRESS } from './constants';
+import { XVS_TOKEN_ID, XVS_VAULT_ADDRESS } from './constants';
+import getBalanceOf, { GetBalanceOfOutput } from '../getBalanceOf';
 
 export interface UseGetXvsVaultPoolBalancesInput {
   stakedTokenAddresses: (string | undefined)[];
@@ -18,24 +18,19 @@ const useGetXvsVaultPoolBalances = ({
 }: UseGetXvsVaultPoolBalancesInput): UseGetXvsVaultPoolBalancesOutput => {
   const web3 = useWeb3();
 
-  // Fetch total amount of tokens staked in each pool
+  // Query total amount of tokens staked for each pool
   const queries: UseQueryOptions<GetBalanceOfOutput>[] = stakedTokenAddresses.map(
     stakedTokenAddress => {
-      const tokenContract = stakedTokenAddress
-        ? getTokenContractByAddress(stakedTokenAddress, web3)
-        : undefined;
+      const tokenContract =
+        stakedTokenAddress && getTokenContractByAddress(stakedTokenAddress, web3);
 
       return {
         queryFn: () =>
           getBalanceOf({
             tokenContract: tokenContract || ({} as Bep20),
-            accountAddress: XVS_VAULT_PROXY_CONTRACT_ADDRESS,
+            accountAddress: XVS_VAULT_ADDRESS,
           }),
-        queryKey: [
-          FunctionKey.GET_BALANCE_OF,
-          stakedTokenAddress,
-          XVS_VAULT_PROXY_CONTRACT_ADDRESS,
-        ],
+        queryKey: [FunctionKey.GET_BALANCE_OF, XVS_TOKEN_ID, XVS_VAULT_ADDRESS],
         enabled: !!tokenContract,
       };
     },
